Move pocket book root helper into common folder
