+++ tests/test_address_book.py
@@ -1,30 +1,10 @@
 import os
-import shutil
-import tempfile
 import unittest
 
 import toml
 
 from pocketbook.address_book import AddressBook
-
-
-class TemporaryPocketBookRoot:
-    def __init__(self):
-        self._root = None
-
-    @property
-    def root(self):
-        return self._root
-
-    def __enter__(self):
-        self._root = tempfile.mkdtemp(prefix='pocketbook_', suffix='_root')
-        return self
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        shutil.rmtree(self._root)
-
-
-SAMPLE_ADDRESS = '2w9uqXLFvCve9TjfoRb2nuztuTx54E1S3AjuB9hWCMMGX5kp4W'
+from .utils import TemporaryPocketBookRoot, SAMPLE_ADDRESS
 
 
 class AddressBookTests(unittest.TestCase):
